Added markdown as content type
the interface EmailContentType was renamed to EmailContent and the method getMessageText was added to it. This way all of the supported mail services will have as input parameter EmailContent, that can be used  to get the right content type AND the exact message that needs to be send. Do it this way using features parsers will be way easier than before